allow for reproducible builds of python packages
See https://reproducible-builds.org/ for why this is good idea
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

Background:
In some distributions like openSUSE, binary rpms contain precompiled .pyc files.

And packages like amqp or twisted dynamically generate .py files at build time
so those have the current time and that timestamp gets embedded
into the .pyc file header.
When we then adapt file timestamps in rpms to be constant,
the timestamp in the .pyc header will no more match
the .py timestamp in the filesystem.
The software will still work, but it will not use the .pyc file as it should.

Original version which doesn't force hashed *.pyc files

Code is originally from gh#296 (never released in this form).

Patch: 0001-allow-for-reproducible-builds-of-python-packages.patch