fix email.generator.py to not replace a non-existent header
This PR replaces GH-1977. The reason for the replacement is
two-fold.

The fix itself is different is that if the CTE header doesn't
exist in the original message, it is inserted. This is important
because the new CTE could be quoted-printable whereas the
original is implicit 8bit.

Also the tests are different. The
test_nonascii_as_string_without_cte test in gh-1977 doesn't
actually test the issue in that it passes without the fix. The
test_nonascii_as_string_without_content_type_and_cte test is
improved here, and even though it doesn't fail without the fix,
it is included for completeness.

Fixed KeyError exception when flattening an email to a string
attempts to replace a non-existent Content-Transfer-Encoding
header.

Code is originally from gh#python/cpython@371146a3f8, it was
released upstream in 3.8.7

Co-authored-by: Mark Sapiro <mark@msapiro.net>
Date: Mon, 19 Oct 2020 16:11:37 -0700
Fixes: bsc#1208443
Patch: bpo27321-email-no-replace-header.patch