[CVE-2023-40217] Check for & avoid the ssl pre-close flaw
Instances of `ssl.SSLSocket` were vulnerable to a bypass of
the TLS handshake and included protections (like certificate
verification) and treating sent unencrypted data as if it were
post-handshake TLS encrypted data.

The vulnerability is caused when a socket is connected, data is
sent by the malicious peer and stored in a buffer, and then the
malicious peer closes the socket within a small timing window
before the other peers’ TLS handshake can begin. After this
sequence of events the closed socket will not immediately attempt
a TLS handshake due to not being connected but will also allow
the buffered data to be read as if a successful TLS handshake had
occurred.

Code is from gh#python/cpython@b4bcc06a9cf, it was released
upstream in 3.8.18.

Co-Authored-By: Gregory P. Smith [Google LLC] <greg@krypto.org>
Fixes: bsc#1214692
Fixes: gh#108315
Fixes: gh#108310
Patch: CVE-2023-40217-avoid-ssl-pre-close.patch